pyproject.toml
[tool.poetry]
name = "yourtube"
version = "0.5.4"
description = "Better youtube recommendations"
authors = ["Filip Sondej <filipsondej@protonmail.com>"]
license = "LGPL-2.1"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
networkx = "^2.6.1"
tqdm = "^4.61.2"
requests = "^2.26.0"
matplotlib = "^3.4.2"
numpy = "^1.21.0"
ipywidgets = "^7.6.3"
ipyevents = "^2.0.1"
voila = "^0.2.10"
magic-wormhole = "^0.12.0"
python-dateutil = "^2.8.2"
krakow = "^0.1.2"
scipy = "^1.7.0"
youtube-transcript-api = "^0.4.1"
pickleDB = "^0.9.2"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"
ipykernel = "^6.0.2"
nb-black = "^1.0.7"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yourtube = 'yourtube.__init__:run'
yourtube-scrape = 'yourtube.scraping:scrape_all_playlists'
yourtube-scrape-watched = 'yourtube.scraping:scrape_watched'
yourtube-scrape-transcripts = 'yourtube.scraping:scrape_transcripts_from_watched_videos'
